requirements.txt
vaderSentiment
nltk
textblob
spacy
pyspellchecker
werkzeug
streamlit